refactor(trafficlights): split process functions out into files, clean up repetition and ts errors
