+++ src/constants.ts
@@ -1,3 +1,5 @@
+import { LightsDataResultStatus, LightsDataValues } from './types';
+
 export const TEST_IDS = {
   go: 'traffic-light-go',
   ready: 'traffic-light-ready',
@@ -7,4 +9,17 @@ export const TEST_IDS = {
   trafficLightValue: 'heywesty-traffic-light-value',
   trafficLightLegend: 'heywesty-traffic-light-legend',
   feedbackMsgContainer: 'heywesty-feedback-msg-container',
+};
+
+const EMPTY_VALUE: LightsDataValues = {
+  title: '',
+  value: '',
+  trend: { color: 'transparent', value: 0 },
+  hasLinks: false,
+};
+
+export const DEFAULT_VALUES = {
+  values: [EMPTY_VALUE],
+  status: LightsDataResultStatus.nodata,
+  invalidThresholds: undefined,
 };
